# backend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Бағдар API",
  "description": "\n**Бағдар** — консультативный автодиспетчер железнодорожного участка.\n\nДемонстрационный прототип на синтетических данных. **Не система управления\nдвижением**: не управляет реальными сигналами, стрелками и поездами и не\nзаменяет устройства СЦБ. Веса, цены и «сэкономленные деньги» условные.\n\nПоток реального времени: `ws /api/stream` (формат — `GET /api/stream/schema`).\n",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/plan": {
   "get": {
    "tags": [
     "planner"
    ],
    "summary": "Действующий, предыдущий или прогнозный план (действующий, сдвинутый на текущие отклонения)",
    "operationId": "get_plan_api_plan_get",
    "parameters": [
     {
      "name": "which",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "pattern": "^(current|previous|projected)$",
       "default": "current",
       "title": "Which"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PlanOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/decisions": {
   "get": {
    "tags": [
     "planner"
    ],
    "summary": "Лента карточек решений",
    "operationId": "get_decisions_api_decisions_get",
    "parameters": [
     {
      "name": "since",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Since"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 300,
       "minimum": 1,
       "default": 100,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DecisionsOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/decisions/{card_id}/action": {
   "post": {
    "tags": [
     "planner"
    ],
    "summary": "Действие диспетчера: отменить решение B или выбрать вариант C",
    "operationId": "decision_action_api_decisions__card_id__action_post",
    "parameters": [
     {
      "name": "card_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Card Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DecisionActionIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ActionResultOut"
        }
       }
      }
     },
     "409": {
      "description": "Окно закрыто, поезд уже на перегоне или вариант недопустим"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/autonomy": {
   "post": {
    "tags": [
     "planner"
    ],
    "summary": "Переключатель «полный авто»: A и B сразу, C — лучшим вариантом",
    "operationId": "set_autonomy_api_autonomy_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AutonomyIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AutonomyConfig"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/index": {
   "get": {
    "tags": [
     "state"
    ],
    "summary": "Индекс эффективности: текущий, прогноз на час по плану, история",
    "operationId": "get_index_api_index_get",
    "parameters": [
     {
      "name": "since",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "Since"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/IndexHistoryOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/traces": {
   "get": {
    "tags": [
     "state"
    ],
    "summary": "Факт движения для графика: точки [t, км] по поездам после since",
    "operationId": "get_traces_api_traces_get",
    "parameters": [
     {
      "name": "since",
      "in": "query",
      "required": false,
      "schema": {
       "type": "number",
       "default": 0.0,
       "title": "Since"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TracesOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/occupancy": {
   "get": {
    "tags": [
     "planner"
    ],
    "summary": "Занятость путей и перегонов для Ганта: факт + действующий или предыдущий план",
    "operationId": "get_occupancy_api_occupancy_get",
    "parameters": [
     {
      "name": "which",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "pattern": "^(current|previous)$",
       "default": "current",
       "title": "Which"
      }
     },
     {
      "name": "t_from",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "T From"
      }
     },
     {
      "name": "t_to",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "T To"
      }
     },
     {
      "name": "at",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "description": "Перемотка: показать занятость на момент at",
       "title": "At"
      },
      "description": "Перемотка: показать занятость на момент at"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OccupancyOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/history/at": {
   "get": {
    "tags": [
     "history"
    ],
    "summary": "Перемотка: состояние, план, лента, карточки и индекс на момент t",
    "operationId": "get_history_at_api_history_at_get",
    "parameters": [
     {
      "name": "t",
      "in": "query",
      "required": true,
      "schema": {
       "type": "number",
       "title": "T"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HistoryAtOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/export/events.csv": {
   "get": {
    "tags": [
     "export"
    ],
    "summary": "CSV: события прогона (из журнала)",
    "operationId": "export_events_api_export_events_csv_get",
    "parameters": [
     {
      "name": "t_from",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "T From"
      }
     },
     {
      "name": "t_to",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "T To"
      }
     },
     {
      "name": "last_min",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number",
         "maximum": 1440,
         "minimum": 1
        },
        {
         "type": "null"
        }
       ],
       "title": "Last Min"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/export/report.pdf": {
   "get": {
    "tags": [
     "export"
    ],
    "summary": "PDF: период, показатели, инциденты, изменения плана, вывод",
    "operationId": "export_pdf_api_export_report_pdf_get",
    "parameters": [
     {
      "name": "t_from",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "T From"
      }
     },
     {
      "name": "t_to",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "T To"
      }
     },
     {
      "name": "last_min",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number",
         "maximum": 1440,
         "minimum": 1
        },
        {
         "type": "null"
        }
       ],
       "title": "Last Min"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/events": {
   "post": {
    "tags": [
     "simulation"
    ],
    "summary": "Внешнее событие или сбой: задержка, закрытие перегона, отказ светофора или стрелки, недоступный путь, ограничение скорости, рост потока, внеочередной поезд",
    "operationId": "post_event_api_events_post",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EventIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EventAck"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "state"
    ],
    "summary": "Журнал событий текущего прогона",
    "operationId": "get_events_api_events_get",
    "parameters": [
     {
      "name": "since",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Since"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 20000,
       "minimum": 1,
       "default": 500,
       "title": "Limit"
      }
     },
     {
      "name": "min_severity",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "debug",
        "info",
        "warn",
        "critical"
       ],
       "type": "string",
       "default": "debug",
       "title": "Min Severity"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EventsOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/incidents/{inc_id}/restore": {
   "post": {
    "tags": [
     "simulation"
    ],
    "summary": "Снять сбой раньше таймера (открыть перегон, починить светофор или стрелку)",
    "operationId": "restore_incident_api_incidents__inc_id__restore_post",
    "parameters": [
     {
      "name": "inc_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Inc Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EventAck"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/sim/control": {
   "post": {
    "tags": [
     "simulation"
    ],
    "summary": "Старт, пауза, ускорение ×1…×100, шаг, сброс",
    "operationId": "sim_control_api_sim_control_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ControlIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ControlOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/sim/load": {
   "post": {
    "tags": [
     "simulation"
    ],
    "summary": "Загрузить сценарий и/или сгенерировать мир с новым seed",
    "operationId": "sim_load_api_sim_load_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LoadIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LoadOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/versus": {
   "get": {
    "tags": [
     "versus"
    ],
    "summary": "Человек против Бағдара: счёт обеих сторон и состояние левой (без Бағдара)",
    "operationId": "get_versus_api_versus_get",
    "parameters": [
     {
      "name": "state",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": true,
       "title": "State"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VersusOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/versus/start": {
   "post": {
    "tags": [
     "versus"
    ],
    "summary": "Запустить сценарий дважды: слева «кто первый пришёл», справа Бағдар",
    "operationId": "start_versus_api_versus_start_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/VersusStartIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VersusOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/versus/hold": {
   "post": {
    "tags": [
     "versus"
    ],
    "summary": "Решение человека слева: придержать поезд на станции",
    "operationId": "versus_hold_api_versus_hold_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/VersusHoldIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EventAck"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "PlanOut": {
    "properties": {
     "version": {
      "type": "integer",
      "title": "Version"
     },
     "created_at": {
      "type": "number",
      "title": "Created At"
     },
     "solver": {
      "type": "string",
      "title": "Solver"
     },
     "status": {
      "type": "string",
      "enum": [
       "feasible",
       "delayed",
       "infeasible"
      ],
      "title": "Status"
     },
     "compute_ms": {
      "type": "number",
      "title": "Compute Ms"
     },
     "notes": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Notes"
     },
     "cost": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Cost"
     },
     "horizon_end": {
      "type": "number",
      "title": "Horizon End"
     },
     "hold_all": {
      "type": "boolean",
      "title": "Hold All"
     },
     "held": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Held"
     },
     "legs": {
      "items": {
       "$ref": "#/components/schemas/PlanLegOut"
      },
      "type": "array",
      "title": "Legs"
     }
    },
    "type": "object",
    "required": [
     "version",
     "created_at",
     "solver",
     "status",
     "compute_ms",
     "notes",
     "cost",
     "horizon_end",
     "hold_all",
     "held",
     "legs"
    ],
    "title": "PlanOut"
   },
...
  }
 }
}